feat: migrate appView to TS
